structure/ clean code - controllers structure modified and module utils
