--- src/controllers/cuack.example.json
@@ -1,7 +1,0 @@
-{
-	"cuack": {
-		"author": "63b6e5f1678d56b86fdb6a08",
-		"content": "This is an example from cuack post",
-		"category": "develop"
-	}
-}
