+++ src/routes/auth.ts
@@ -1,23 +1,26 @@
 import { Router } from 'express';
 import { check } from 'express-validator';
-import { login, googleSignIn } from '../controllers/authController';
+import { login, googleSignIn } from '../controllers/auth';
 import { validateFields } from '../middlewares/validateField';
 
 const router = Router();
 
 router.post(
-	'/',
-	[
-		check('email', 'Please, send a valid email.').isEmail(),
-		check('password', 'The password is required.').not().isEmpty(),
-		validateFields,
-	],
-	login
+  '/',
+  [
+    check('email', 'Please, send a valid email.').isEmail(),
+    check('password', 'The password is required.').not().isEmpty(),
+    validateFields,
+  ],
+  login
 );
 
 router.post(
-	'/google',
-	[check('id_token', 'Token de Google es requerido').not().isEmpty(), validateFields],
-	googleSignIn
+  '/google',
+  [
+    check('id_token', 'Token de Google es requerido').not().isEmpty(),
+    validateFields,
+  ],
+  googleSignIn
 );
 export default router;
